+++ crates/egui_pull_to_refresh/examples/pull_refresh_scroll_area.rs
@@ -26,6 +26,9 @@ pub fn main() -> eframe::Result<()> {
         Default::default(),
         move |ctx, _frame| {
             CentralPanel::default().show(ctx, |ui| {
+                // Disable text selection, so it doesn't interfere with the drag gesture
+                ui.style_mut().interaction.selectable_labels = false;
+                ui.style_mut().interaction.multi_widget_text_select = false;
                 let current_state = state.lock().unwrap().clone();
 
                 let response = PullToRefresh::new(current_state.loading).scroll_area_ui(ui, |ui| {