+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # hello_egui changelog
 
+## 0.3.0
+- update all crates for egui 0.26 support
+
 ## 0.2.0
 - Removed all crates from the default features, and instead added a new feature `all` that includes all crates
 - Updated egui to 0.25, dropping support for older versions