+++ crates/egui_infinite_scroll/CHANGELOG.md
@@ -1,4 +1,7 @@
 # egui_infinite_scroll changelog
 
+## 0.2.0
+- update egui to 0.26
+
 ## 0.1.0 
 Initial release