+++ crates/egui_pull_to_refresh/CHANGELOG.md
@@ -1,4 +1,9 @@
 # egui_pull_to_refresh changelog
 
+## 0.3.0
+- Updated egui to 0.26
+- Please note that you might want to disable egui's new drag selection feature
+where you use PullToRefresh, so it doesn't interfere with the drag gesture.
+
 ## 0.2.0
 - Updated egui to 0.25, dropping support for older versions