+++ crates/egui_suspense/CHANGELOG.md
@@ -1,5 +1,8 @@
 # egui_suspense changelog
 
+## 0.3.0
+- Updated egui to 0.26
+
 ## 0.2.0
 - Updated egui to 0.25, dropping support for older versions
 - Added async version of the api