+++ crates/egui_thumbhash/CHANGELOG.md
@@ -1,4 +1,7 @@
 # egui_thumbhash changelog
 
+## 0.2.0
+- Updated egui to 0.26
+
 ## 0.1.0
 Initial release