+++ crates/hello_egui_utils/CHANGELOG.md
@@ -1,5 +1,8 @@
 # hello_egui_utils changelog
 
+## 0.3.0
+- updated egui to 0.26
+
 ## 0.2.0
 - Added async utilities
 - Updated egui to 0.25, dropping support for older versions